+++ peekingduck/pipeline/nodes/abstract_node.py
@@ -22,7 +22,7 @@ from abc import ABCMeta, abstractmethod
 from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
-from peekingduck.configloader import ConfigLoader
+from peekingduck.config_loader import ConfigLoader
 from peekingduck.utils.create_node_helper import obj_det_change_class_name_to_id
