+++ tests/loaders/test_declarative_loader.py
